+++ src/index.js
@@ -13,21 +13,14 @@ import {
   Home,  
   Setup
 } from "./pages";
-import { 
-  Footers, 
-  NavigationBar 
-} from "./components";
-
 ReactDOM.render(
   <Router>
-    <NavigationBar />
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/generate" element={<Generate />} />
         <Route path="/setup" element={<Setup />} />
         <Route path="/history" element={<History />} />
       </Routes>
-    <Footers />
   </Router>,
 
   document.getElementById("root")
