--- src/pages/Generate.js
@@ -1,15 +1,0 @@
-import Container from "react-bootstrap/esm/Container";
-
-function Generate() {
-  return (
-    <div className="text-center mt-5">
-      <Container className="pt-5">
-        <div className="bg-white border border-secondary py-3">
-          <h1>Generate</h1>
-        </div>
-      </Container>
-    </div>
-  );
-}
-
-export default Generate;
