+++ src/pages/History/index.js
@@ -1,0 +1,21 @@
+import Container from "react-bootstrap/esm/Container";
+import { 
+    Footers, 
+    NavigationBar 
+} from "../../components";
+
+function History() {
+    return (
+      <div className="text-center mt-5">
+        <NavigationBar />
+        <Container className="pt-5">
+            <div className="bg-white border border-secondary py-3">
+            <h1>History</h1>
+            </div>
+        </Container>
+        <Footers />
+    </div>
+    );
+  }
+  
+  export default History;
