--- src/pages/Setup.js
@@ -1,15 +1,0 @@
-import Container from "react-bootstrap/esm/Container";
-
-function Setup() {
-  return (
-    <div className="text-center mt-5">
-      <Container className="pt-5">
-        <div className="bg-white border border-secondary py-3">
-          <h1>Setup</h1>
-        </div>
-      </Container>
-    </div>
-  );
-}
-
-export default Setup;
